Reorganize unit tests for react-wildcat-prefetch.
