+++ packages/react-wildcat-prefetch/test/browser/fixtures/HelloES7.js
